test(battle): merge Phase5 integration tests into BattleActionExecutor.test; remove redundant file
